+++ src/one/methods/compute/freq.js
@@ -4,7 +4,7 @@ const freq = function (docs) {
   for (let i = 0; i < docs.length; i += 1) {
     for (let t = 0; t < docs[i].length; t += 1) {
       let term = docs[i][t]
-      let word = term.implicit || term.normal
+      let word = term.machine || term.normal
       counts[word] = counts[word] || 0
       counts[word] += 1
     }
@@ -13,7 +13,7 @@ const freq = function (docs) {
   for (let i = 0; i < docs.length; i += 1) {
     for (let t = 0; t < docs[i].length; t += 1) {
       let term = docs[i][t]
-      let word = term.implicit || term.normal
+      let word = term.machine || term.normal
       term.freq = counts[word]
     }
   }