+++ src/one/methods/compute/index.js
@@ -2,13 +2,7 @@ import normalize from './normal/index.js'
 import alias from './alias.js'
 import offset from './offset.js'
 import freq from './freq.js'
-
-const machine = function (term) {
-  let str = term.implicit || term.normal || term.text
-  str = str.replace(/'s$/, '') //no apostrophes
-  str = str.replace(/[-]/, '') //no hyphens
-  term.machine = str
-}
+import machine from './machine.js'
 
 const wordCount = function (docs) {
   let n = 0
@@ -35,10 +29,10 @@ const termLoop = function (docs, method, model) {
 const compute = {
   // text formats
   normal: docs => termLoop(docs, normalize),
-  // like 'normal', but no hyphens or apostrophes
-  machine: docs => termLoop(docs, machine),
   // known word swaps
   alias: (docs, model) => termLoop(docs, alias, model),
+  // like 'normal', but no hyphens or apostrophes
+  machine: docs => termLoop(docs, machine),
 
   // count repeated words
   freq,